requirements.txt
numpy>=1.23,<2.0
tensorflow>=2.12
matplotlib>=3.7
scikit-learn>=1.3
streamlit>=1.28
pillow>=7.1.0

# Nota: En macOS con Apple Silicon (M1/M2/M3), se recomienda instalar:
# tensorflow-macos==2.15.0
# tensorflow-metal==1.1.0
# En lugar de tensorflow genérico

